Two string games - investigating crashes & infinite loops

+++ two_strings_game/hackerrank.cpp
@@ -183,7 +183,7 @@ bool isWinningPosition(const string & A, const string & B, const Position & pos)
     return loosingPos ? false : true;
 }
 
-void printOutput(Position & pos, int k)
+void printOutput(Position & pos, unsigned long long int k)
 {
     // cout << "k = " << k << ", position = " << pos.first << " , " << pos.second << endl;
     cout << pos.first << endl;
@@ -195,21 +195,21 @@ int main() {
 
     string tempString;
     getline(cin, tempString, ' ');
-        int N = stoi(tempString);
+    unsigned int N = stoi(tempString);
     getline(cin, tempString, ' ');
-    int M = stoi(tempString);
+    unsigned int M = stoi(tempString);
     getline(cin, tempString);
-    int K = stoi(tempString);
+    unsigned long long int K = stoll(tempString);
 
     string A, B;
     getline(cin, A);
     getline(cin, B);
 
-        SubstringGenerator generatorA(A);
-        SubstringGenerator generatorB(B);
+    SubstringGenerator generatorA(A);
+    SubstringGenerator generatorB(B);
 
     Position position;
-    int k = 0;
+    unsigned long long int k = 0;
 
     string substringA;
     string substringB;
